Revert "Redundant codes"
This reverts commit b007188519fb9ed49f5a70a1d77d4ae033987544.

# Conflicts:
#	.vs/Ultrapowa Clash Server/v14/.suo

Reason for Revert:

No Need to change the whole Code, Berkan you have to work at MySQL improvements, no Core changes.

+++ Ultrapowa Clash Server/Helpers/GamePlayUtil.cs
@@ -1,6 +1,15 @@
 ﻿using System;
-using UCS.Core;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.IO;
+using System.Collections.Concurrent;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
 using UCS.GameFiles;
+using UCS.Core;
+using UCS.Logic;
 
 namespace UCS.Helpers
 {
@@ -20,12 +29,13 @@ namespace UCS.Helpers
 
         public static int CalculateResourceCost(int sup, int inf, int supCost, int infCost, int amount)
         {
-            return (int) Math.Round((supCost - infCost)*(long) (amount - inf)/(sup - inf*1.0)) + infCost;
+            return (int)Math.Round((long)(supCost - infCost) * (long)(amount - inf) / (sup - inf * 1.0)) + infCost;
         }
 
         public static int CalculateSpeedUpCost(int sup, int inf, int supCost, int infCost, int amount)
         {
-            return (int) Math.Round((supCost - infCost)*(long) (amount - inf)/(sup - inf*1.0)) + infCost;
+            return (int)Math.Round((long)(supCost - infCost) * (long)(amount - inf) / (sup - inf * 1.0)) + infCost;
         }
     }
+
 }
